Basic linter for Java
Summary:
This is a simple script&regex linter that calls the gradle scripts. As of now, what it does is:

 - The linter fails when the gradle build fails (tested for: syntax error, unit test error; untested for static code analysis)
 - The linter succeeds when the gradle build succeeds
 - The linter ignores lines that contain "warning" (because there are currently some warnings emitted by the javadoc generator, see T221)

As of now, the linter executes:

 - Compilation
 - Unit testing
 - Static code analysis
 - Exportation of Javadoc

As of now, the linter is not able to detail what the problem is. It is only able to execute/fail on the afore-mentioned cases. If it fails, you have to call the Gradle script yourself to get more details.

Test Plan: Call 'arc lint' and check that any error in the code is reported.

Reviewers: Info-Screen, WyldBot

Reviewed By: Info-Screen

Subscribers: WyldBot, Info-Screen

Maniphest Tasks: T215

Differential Revision: https://phabricator.wildfyre.net/D167